Update job_wrapper.py
Added command syntax for remote-run.

Also suggest an update to compute.yml setup to auto generate on the fly based on environment variables that can be passed to the azure container instance at runtime and which can be set by the user in the application settings. This gives the user to change (say) the number of CPUs to use via the application instead of uploading a new compute.yml.