feat: move plugins to __one__ folder
There is no destinction between 'app plugins' and 'ui plugin.
There are only packages that exports multiple plugins, of either 'page'
or 'ui' type. The file structure should reflect this.

Also moved 'dmt/common' out of 'plugins' and renamed it to 'dmt-core'.
This will later be published as a standalone NPM package.